Refactor

+++ src/com/laoapps/handler/DepartmentHandler.java
@@ -1,5 +1,6 @@
 package com.laoapps.handler;
 
+import com.google.common.base.Strings;
 import com.google.gson.Gson;
 import com.google.gson.JsonObject;
 import com.laoapps.database.connector.CustomInterceptor;
@@ -8,9 +9,9 @@ import com.laoapps.database.entity.Department;
 import com.laoapps.models.CheckJwtResult;
 import com.laoapps.utils.MyCommon;
 import com.laoapps.utils.Naming;
-import com.laoapps.websocker.response.Response;
-import com.laoapps.websocker.response.ResponseData;
-import com.laoapps.websocker.response.ResponseBody;
+import com.laoapps.socket.response.Response;
+import com.laoapps.socket.response.ResponseData;
+import com.laoapps.socket.response.ResponseBody;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
 
@@ -35,20 +36,20 @@ public class DepartmentHandler {
 
     public String create(JsonObject data, CheckJwtResult checkJwtResult) {
 
-        try (Session session = factory.withOptions().interceptor(new CustomInterceptor(checkJwtResult.getCheckJwt().getParent())).openSession()) {
+        String name = data.get(Naming.NAME).getAsString();
+        String parent = null, remark = null;
 
-            String name = data.get(Naming.NAME).getAsString();
-            String parent = data.get(Naming.PARENT).getAsString();
-            String remark = null;
+        if (data.has(Naming.PARENT)) parent = data.get(Naming.PARENT).getAsString();
+        if (data.has(Naming.REMARK)) remark = data.get(Naming.REMARK).getAsString();
 
-            if (data.has(Naming.REMARK)) remark = data.get(Naming.REMARK).getAsString();
+        try (Session session = factory.withOptions().interceptor(new CustomInterceptor(checkJwtResult.getCheckJwt().getCompany())).openSession()) {
 
             session.beginTransaction();
 
-            String role = (String) session.createQuery("select role from Users where uuid = '" + checkJwtResult.getCheckJwt().getUuid() + "'").uniqueResult();
+            String role = (String) session.createQuery("select role from Profiles where uuid = '" + checkJwtResult.getCheckJwt().getUuid() + "'").uniqueResult();
 
             if (!role.equals(Naming.admin) && !role.equals(Naming.hr)) {
-                Response response = new Response(new ResponseBody(Naming.department, Naming.create, Naming.fail, "Access denied", null));
+                Response response = new Response(new ResponseBody(Naming.department, Naming.create, Naming.fail, "access denied", null));
                 MyCommon.printMessage(response.toString());
 
                 return gson.toJson(response);
@@ -57,8 +58,8 @@ public class DepartmentHandler {
             Department department = new Department();
 
             department.setName(name);
-            department.setParent(parent);
-            if (remark != null) department.setRemark(remark);
+            if (!Strings.isNullOrEmpty(parent)) department.setParent(parent); else department.setParent(Naming.self);
+            if (!Strings.isNullOrEmpty(remark)) department.setRemark(remark);
             department.setCreatedAt(MyCommon.currentTime());
             department.setUuid(MyCommon.generateUuid());
 
@@ -85,9 +86,39 @@ public class DepartmentHandler {
 
     }
 
+    public String getById(JsonObject data, CheckJwtResult checkJwtResult) {
+        try (Session session = factory.withOptions().interceptor(new CustomInterceptor(checkJwtResult.getCheckJwt().getCompany())).openSession()) {
+
+            int id = data.get(Naming.ID).getAsInt();
+
+            session.beginTransaction();
+
+            Department department = (Department) session.createQuery("from Department where id = " + id + "").uniqueResult();
+
+            if (department == null) throw new NullPointerException("Department id: " + id + " not exists");
+
+            session.getTransaction().commit();
+
+            ResponseData responseData = new ResponseData();
+            responseData.setDepartment(department);
+            responseData.setJwt(checkJwtResult.getCheckJwt().getJwt());
+
+            Response response = new Response(new ResponseBody(Naming.department, Naming.getById, Naming.success, "successful", responseData));
+            MyCommon.printMessage(response.toString());
+            return gson.toJson(response);
+
+        } catch (Exception e) {
+            e.printStackTrace();
+            Response response = new Response(new ResponseBody(Naming.department, Naming.getById, Naming.fail, e.getMessage(), null));
+            MyCommon.printMessage(response.toString());
+
+            return gson.toJson(response);
+        }
+    }
+
     public String get(JsonObject data, CheckJwtResult checkJwtResult) {
 
-        try (Session session = factory.withOptions().interceptor(new CustomInterceptor(checkJwtResult.getCheckJwt().getParent())).openSession()) {
+        try (Session session = factory.withOptions().interceptor(new CustomInterceptor(checkJwtResult.getCheckJwt().getCompany())).openSession()) {
 
             int page = data.get(Naming.page).getAsInt();
             int limit = data.get(Naming.limit).getAsInt();
@@ -98,18 +129,19 @@ public class DepartmentHandler {
 
             session.beginTransaction();
 
-            String searchQuery = "from Department where id = " + keywordId + " or name like '" + keyword +
-                    "' or parent like '" + keyword + "' or remark like '" + keyword + "' or createdAt like '" + keyword +
-                    "' or updatedAt like '" + keyword + "' or uuid like '" + keyword + "'";
+            String searchQuery = "from Department where id = " + keywordId + " or name like '%" + keyword +
+                    "%' or parent like '%" + keyword + "%' or remark like '%" + keyword + "%' or createdAt like '%" + keyword +
+                    "%' or updatedAt like '%" + keyword + "%' or uuid like '%" + keyword + "%' order by id desc";
 
-            String query = (keyword == null) ? "from Department" : searchQuery;
+            String query = (keyword == null) ? "from Department order by id desc" : searchQuery;
 
             String countQuery = (keyword == null) ? "select count(id) from Department" : "select count(id) " + searchQuery;
             Long count = (Long) session.createQuery(countQuery).uniqueResult();
             Integer totalPage = (count.intValue() + limit - 1) / limit;
 
             @SuppressWarnings("unchecked")
-            ArrayList<Department> departments = (ArrayList<Department>) session.createQuery(query).setFirstResult((page - 1) * limit).setMaxResults(limit).getResultList();
+            ArrayList<Department> departments = (ArrayList<Department>) session.createQuery(query).setFirstResult((page - 1) * limit)
+                    .setMaxResults(limit).getResultList();
 
             session.getTransaction().commit();
 
@@ -137,11 +169,17 @@ public class DepartmentHandler {
 
     public String update(JsonObject data, CheckJwtResult checkJwtResult) {
 
-        try (Session session = factory.withOptions().interceptor(new CustomInterceptor(checkJwtResult.getCheckJwt().getParent())).openSession()) {
+        try (Session session = factory.withOptions().interceptor(new CustomInterceptor(checkJwtResult.getCheckJwt().getCompany())).openSession()) {
+
+            String name = null, parent = null, remark = null;
+
+            if (data.has(Naming.NAME)) name = data.get(Naming.NAME).getAsString();
+            if (data.has(Naming.PARENT)) parent = data.get(Naming.PARENT).getAsString();
+            if (data.has(Naming.REMARK)) remark = data.get(Naming.REMARK).getAsString();
 
             session.beginTransaction();
 
-            String role = (String) session.createQuery("select role from Users where uuid = '" + checkJwtResult.getCheckJwt().getUuid() + "'").uniqueResult();
+            String role = (String) session.createQuery("select role from Profiles where uuid = '" + checkJwtResult.getCheckJwt().getUuid() + "'").uniqueResult();
 
             if (!role.equals(Naming.admin) && !role.equals(Naming.hr)) {
                 Response response = new Response(new ResponseBody(Naming.department, Naming.update, Naming.fail, "access denied", null));
@@ -152,9 +190,9 @@ public class DepartmentHandler {
 
             Department department = session.get(Department.class, data.get(Naming.UUID).getAsString());
 
-            if (data.has(Naming.NAME)) department.setName(data.get(Naming.NAME).getAsString());
-            if (data.has(Naming.PARENT)) department.setParent(data.get(Naming.PARENT).getAsString());
-            if (data.has(Naming.REMARK)) department.setRemark(data.get(Naming.REMARK).getAsString());
+            if (!Strings.isNullOrEmpty(name)) department.setName(name);
+            if (!Strings.isNullOrEmpty(parent)) department.setParent(parent); else department.setParent(Naming.self);
+            if (!Strings.isNullOrEmpty(remark)) department.setRemark(remark);
             department.setUpdatedAt(MyCommon.currentTime());
 
             session.update(department);
@@ -181,11 +219,13 @@ public class DepartmentHandler {
 
     public String delete(JsonObject data, CheckJwtResult checkJwtResult) {
 
-        try (Session session = factory.withOptions().interceptor(new CustomInterceptor(checkJwtResult.getCheckJwt().getParent())).openSession()) {
+        String uuid = data.get(Naming.UUID).getAsString();
+
+        try (Session session = factory.withOptions().interceptor(new CustomInterceptor(checkJwtResult.getCheckJwt().getCompany())).openSession()) {
 
             session.beginTransaction();
 
-            String role = (String) session.createQuery("select role from Users where uuid = '" + checkJwtResult.getCheckJwt().getUuid() + "'").uniqueResult();
+            String role = (String) session.createQuery("select role from Profiles where uuid = '" + checkJwtResult.getCheckJwt().getUuid() + "'").uniqueResult();
 
             if (!role.equals(Naming.admin) && !role.equals(Naming.hr)) {
                 Response response = new Response(new ResponseBody(Naming.department, Naming.delete, Naming.fail, "Access denied", null));
@@ -194,7 +234,9 @@ public class DepartmentHandler {
                 return gson.toJson(response);
             }
 
-            Department department = session.get(Department.class, data.get(Naming.UUID).getAsString());
+            Department department = session.get(Department.class, uuid);
+
+            if (department == null) throw new NullPointerException("Department uuid: " + uuid + " not exists");
 
             session.delete(department);
